update, after merge

+++ src/main/java/com/example/video_chat/repository/FileRepository.java
@@ -1,8 +1,8 @@
 package com.example.video_chat.repository;
-import com.example.video_chat.repository.entity.FileEntity;
+import com.example.video_chat.domain.entities.FileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileRepository extends JpaRepository<FileEntity, Integer> {
+public interface FileRepository extends JpaRepository<FileEntity, Long> {
     void delete(FileEntity fileEntity);
     FileEntity findByUrl(String url);
 }
